feat(explore): add Zipkin source to Trace inspect
- Source toggle now switches the cross-layer Trace inspect between Native
  and Zipkin. Zipkin picks a service from the global Zipkin index (no
  layer/id), with optional remote-service + span filters and an
  annotation-query condition; native-only conditions (status/order/tags)
  are hidden.
- BFF runs the full Zipkin REST query (remoteService/span/annotation/
  duration/window/limit) against oap.zipkinUrl — not the GraphQL port,
  which 404s. Soft-fails to an unreachable envelope.
- Extract the Zipkin span waterfall + detail into a shared
  ZipkinTraceDetailCard reused by the per-layer Zipkin tab and inspect;
  Zipkin rows adapt onto the shared trace list + distribution.
